Feature: Simplify polars and pandas
