+++ dfd/dataset/__init__.py
@@ -1,11 +1,11 @@
 """Tabular dataset analysis public API."""
 
 from .analyses import (
-    PandasTabularAnalyses,
-    PolarsTabularAnalyses,
     TabularDataContext,
     TabularStatistics,
 )
+from .pandas_strategy import PandasTabularAnalyses
+from .polars_strategy import PolarsTabularAnalyses
 
 __all__ = [
     'PandasTabularAnalyses',
