+++ tests/test_create.py
@@ -2,7 +2,9 @@ import pandas as pd
 import polars as pl
 
 from dfd.create import Datasheet
-from dfd.dataset.analyses import PandasTabularAnalyses, PolarsTabularAnalyses, TabularStatistics
+from dfd.dataset.analyses import TabularStatistics
+from dfd.dataset.pandas_strategy import PandasTabularAnalyses
+from dfd.dataset.polars_strategy import PolarsTabularAnalyses
 
 # create test data
 d = {'a': [1, 2], 'b': ['A', 'B']}
